# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API des Commentaires",
        "version": "1.0"
    },
    "paths": {
...
        "/api/commentaires": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Lister tous les commentaires",
                "operationId": "ee4d918480e497cee636424744ecfebb",
                "responses": {
                    "200": {
                        "description": "Liste des commentaires récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Commentaire"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Créer un nouveau commentaire",
                "operationId": "983601800620fa4f84fc84a6fc5dc405",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCommentaireRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commentaire"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/commentaires/{id}": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Afficher un commentaire spécifique",
                "operationId": "e8498196bc11759bbebadfd81b0d7484",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire récupéré avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commentaire"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commentaire non trouvé."
                    }
                }
            },
            "put": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Mettre à jour un commentaire spécifique",
                "operationId": "e72e2de188dc869574a02ac882e99543",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCommentaireRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commentaire mis à jour avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commentaire"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Action non autorisée."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprimer un commentaire spécifique",
                "operationId": "c4030e179dc97d2c0a895705c82bfa13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Commentaire supprimé avec succès."
                    },
                    "403": {
                        "description": "Action non autorisée."
                    }
                }
            }
        },
        "/api/domaines-sportifs": {
            "get": {
                "tags": [
                    "Domaines Sportifs"
                ],
                "summary": "Lister tous les domaines sportifs",
                "operationId": "ebc73829886412a28d7e19bd33a58cae",
                "responses": {
                    "200": {
                        "description": "Liste des domaines sportifs récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DomaineSportif"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Domaines Sportifs"
                ],
                "summary": "Créer un nouveau domaine sportif",
                "operationId": "c65473a2c9ff3b186ee3302adc9d3517",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDomaineSportifRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Domaine sportif créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomaineSportif"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/domaines-sportifs/{id}": {
            "get": {
                "tags": [
                    "Domaines Sportifs"
                ],
                "summary": "Afficher un domaine sportif spécifique",
                "operationId": "ad699c8a73acfa9c75d052c40deee0a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Domaine sportif récupéré avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomaineSportif"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Domaine sportif non trouvé."
                    }
                }
            },
            "put": {
                "tags": [
                    "Domaines Sportifs"
                ],
                "summary": "Mettre à jour un domaine sportif spécifique",
                "operationId": "3d3b309b4f02dfa16c1f98310c2ff0df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDomaineSportifRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Domaine sportif mis à jour avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomaineSportif"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Domaine sportif non trouvé."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Domaines Sportifs"
                ],
                "summary": "Supprimer un domaine sportif spécifique",
                "operationId": "b64c5fff91182dc8511917509a1b1f89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Domaine sportif supprimé avec succès."
                    },
                    "404": {
                        "description": "Domaine sportif non trouvé."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Commentaire": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "Ceci est un commentaire."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreCommentaireRequest": {
                "required": [
                    "content"
                ],
                "properties": {
                    "content": {
                        "type": "string",
                        "example": "Ceci est un commentaire."
                    }
                },
                "type": "object"
            },
            "UpdateCommentaireRequest": {
                "required": [
                    "content"
                ],
                "properties": {
                    "content": {
                        "type": "string",
                        "example": "Ceci est un commentaire modifié."
                    }
                },
                "type": "object"
            },
            "DomaineSportif": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Football"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "Category": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Catégorie de Football"
                    }
                },
                "type": "object"
            },
            "StoreDomaineSportifRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Football"
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateDomaineSportifRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Football Modifié"
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}